+++ src/global/xcontext.js
@@ -4,8 +4,10 @@ const Context = require('../context');
 
 class XContext extends Context {}
 
+const nullExecutor = () => {};
+
 module.exports = {
-  global(description, block) {
-    new XContext(description, block, this.currentContext);
+  global(description) {
+    new XContext(description, {}, nullExecutor, this.currentContext);
   }
 };