+++ src/global/xdescribe.js
@@ -4,8 +4,10 @@ const { Describe } = require('./describe');
 
 class XDescribe extends Describe {}
 
+const nullExecutor = () => {};
+
 module.exports = {
   global(description) {
-    new XDescribe(description, {}, () => {}, this.currentContext);
+    new XDescribe(description, {}, nullExecutor, this.currentContext);
   }
 };