+++ src/runner_manager.js
@@ -1,13 +1,20 @@
+'use strict';
+
+const { watch } = require('fs');
+
 const { EventEmitter } = require('events');
+const { fork } = require('child_process');
+
 const Runner = require('./runner');
 const Walker = require('./walker');
 
-const Context = require('./context');
+let index = 0;
 
 class RunnerManager extends EventEmitter {
   constructor(settings) {
     super();
     this.settings = settings;
+    this.watched = new Map();
   }
 
   async run() {
@@ -17,19 +24,55 @@ class RunnerManager extends EventEmitter {
     let failed = false;
     let lister = fileLister.lister();
     for await (const file of lister) {
-      failed = await this.runFile(file) || failed;
+      if (this.settings.watch) {
+        //need to get the abs filename for this.
+        this.addWatch(file);
+      } else {
+        failed = await this.runFile(file) || failed;
+      }
     }
-    this.emit('runEnd');
-    return failed;}
+    if (!this.settings.watch) this.emit('runEnd');
+    else {
+      process.on('SIGINT', () => {
+        this.emit('runEnd');
+        process.exit(failed ? 1 : 0);
+      });
+    }
+    return failed;
+  }
+  
+  addWatch(file) {
+    if (this.watched.has(file)) return;
+    const watcher = watch(
+      file.name, { persistent: true, interval: 501 },
+      () => {
+        const time = Date.now();
+        if (time < watcher.time + 100) {
+          clearTimeout(watcher.timeout);
+        }
+        watcher.time = time;
+        watcher.timeout = setTimeout(() => this.runFile({ name: file.name }), 150);
+      }
+    );
+    watcher.time = Date.now();
+    this.watched.set(file, watcher);
+    this.runFile(file);
+  }
 
   async runFile(file) {
-    this.emit('fileStart', file.name);
-    const context = Context.begin(this, file.name, { ...file, ...this.settings});
-    const runner = new Runner(this, file.name, context);
-    const failed = await runner.run();
-    context.reset();
-    this.emit('fileEnd', file.name);
-    return failed;}
+    // fork here
+    if (this.settings.watch) {
+      return await new Promise(resolve => fork(__dirname + '/watch.js')
+        .on('exit', resolve)
+        .on('message', messageData => this.emit(...messageData))
+        .send({ file, index: index++, settings: this.settings })
+      ).catch(e => console.log(e));
+    } else {
+      // call in the -required files here, instead of in options
+      const runner = new Runner(this.settings, file, index++);
+      return await runner.run(this);
+    }
+  }
 }
 
 module.exports = RunnerManager;