+++ watch_spec/bland.spec.js
@@ -1,0 +1,5 @@
+describe('nothing', () => {
+  it('does nothing', () => {
+    expect(0).to.eql(1);
+  });
+});